refactor: webapp configuration reorganization
