+++ src/components/ScenarioParameters/README.md
@@ -14,7 +14,7 @@ const scenarioParametersTabs = [
 ];
 ```
 
-- Open <path_to_your_app>/src/configs/ScenarioParametersTabs.config.js
+- Open <path_to_your_app>/src/config/ScenarioParameters.js
 - Add the config associated to your new tab
 
 ```
@@ -27,9 +27,9 @@ const scenarioParametersTabs = [
   runTemplateIds: ['2', '3']                            // The template IDs that have to display your component
 }
 ```
-**Note #1:** if you inserted your component in an existing ones, don't forget to update the keys of the other components in `scenarioParametersTabs`, as well as the ids in the ScenarioParametersTabs.config.js file, accordingly.
+**Note #1:** if you inserted your component in an existing ones, don't forget to update the keys of the other components in `scenarioParametersTabs`, as well as the ids in the src/config/ScenarioParameters.js file, accordingly.
 
-**Note #2:** *both* `key` from `<MyComponent>` *and* `id` in ScenarioParametersTabs.config.js must have the same value (here 42).
+**Note #2:** *both* `key` from `<MyComponent>` *and* `id` in src/config/ScenarioParameters.js must have the same value (here 42).
 
 **Note #3:** If you want to add a translation, complete every translation file as the following example:
 ```
@@ -50,7 +50,7 @@ const scenarioParametersTabs = [
 # Remove a tab from Scenario Parameters
 - Open <path_to_your_app>/src/components/ScenarioParameters/ScenarioParameters.js
 - Remove your component
-- Open <path_to_your_app>/src/configs/ScenarioParametersTabs.config.js
+- Open <path_to_your_app>/src/config/ScenarioParameters.js
 - Remove the config of your tab
 - Clean imports in ScenarioParameters, and the code no longer needed
 - Remove the keys from the translation files no longer used
