+++ src/config/AppInstance.js
@@ -1,0 +1,14 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// Azure App registration details
+export const AZURE_TENANT_ID = 'e413b834-8be8-4822-a370-be619545cb49';
+export const APP_REGISTRATION_CLIENT_ID = 'd104fbd6-9464-45d5-a022-83c90ad56906';
+export const APPLICATION_INSIGHTS_INSTRUMENTATION_KEY = 'f198a852-15d8-48da-9cd0-1e6d18708489';
+
+// API information
+export const DEFAULT_BASE_PATH = 'https://dev.api.cosmotech.com';
+
+// Cosmo solution identifiers
+export const ORGANIZATION_ID = 'O-gZYpnd27G7';
+export const WORKSPACE_ID = 'W-rXeBwRa0PM';
