+++ src/config/README.md
@@ -1,0 +1,9 @@
+# Webapp configuration
+
+This folder contains several configuration files for the WebApp:
+* **appInstance.js** defines constants specific to your application instance: you will need to change these values when
+  creating a new staging or prod version of your application
+* **appConfiguration.js** defines constants to customize the application behavior: these values are usually common for all
+  instances of a given solution
+* **dashboards.js** regroups the definition of all PowerBI reports of your application
+* **scenarioParameters.js** defines the configuration of the scenario parameters, that are specific to your solution
