+++ src/services/config/Api.js
@@ -8,12 +8,8 @@ import {
   SolutionApiFactory,
   WorkspaceApiFactory
 } from '@cosmotech/api-ts';
-import { clientApi } from '../services/ClientApi';
-
-const DEFAULT_BASE_PATH = 'https://dev.api.cosmotech.com';
-
-// Polling delay to update running scenario status (seconds)
-const SCENARIO_STATUS_POLLING_DELAY = 10000;
+import { clientApi } from '../ClientApi';
+import { DEFAULT_BASE_PATH } from '../../config/AppInstance';
 
 export const Api = {
   Scenarios: ScenarioApiFactory(null, DEFAULT_BASE_PATH, clientApi),
@@ -22,10 +18,3 @@ export const Api = {
   Datasets: DatasetApiFactory(null, DEFAULT_BASE_PATH, clientApi),
   Workspaces: WorkspaceApiFactory(null, DEFAULT_BASE_PATH, clientApi)
 };
-
-// REST API data object
-const API_CONFIG = {
-  scenarioStatusPollingDelay: SCENARIO_STATUS_POLLING_DELAY
-};
-
-export default API_CONFIG;
