+++ src/services/config/Auth.js
@@ -3,9 +3,18 @@
 
 import { Auth, AuthDev } from '@cosmotech/core';
 import { AuthMSAL } from '@cosmotech/azure';
+import { AZURE_TENANT_ID, APP_REGISTRATION_CLIENT_ID, DEFAULT_BASE_PATH } from '../../config/AppInstance.js';
 
-const AZURE_CLIENT_ID = 'd104fbd6-9464-45d5-a022-83c90ad56906';
-const AZURE_TENANT_ID = 'e413b834-8be8-4822-a370-be619545cb49';
+// Expected input base path format: https://dev.api.cosmotech.com or https://dev.api.cosmotech.com/
+// Generated output: 'http://dev.api.cosmotech.com/platform'
+function forgeScopeFromBasePath (basePath) {
+  const domainNameRegexp = /^https:\/\/(.*)(\/)?$/;
+  const domainName = basePath.replace(/\/$/, '').match(domainNameRegexp);
+  if (!domainName) {
+    throw new Error('DEFAULT_BASE_PATH does not match the expected format (e.g. https://dev.api.cosmotech.com)');
+  }
+  return `http://${domainName[1]}/platform`;
+}
 
 // AuthMSAL configuration
 const MSAL_CONFIG = {
@@ -13,11 +22,11 @@ const MSAL_CONFIG = {
     scopes: ['user.read']
   },
   accessRequest: {
-    scopes: ['http://dev.api.cosmotech.com/platform']
+    scopes: [forgeScopeFromBasePath(DEFAULT_BASE_PATH)]
   },
   msalConfig: {
     auth: {
-      clientId: AZURE_CLIENT_ID,
+      clientId: APP_REGISTRATION_CLIENT_ID,
       redirectUri: window.location.protocol + '//' + window.location.host + '/scenario',
       authority: `https://login.microsoftonline.com/${AZURE_TENANT_ID}`,
       knownAuthorities: [`https://login.microsoftonline.com/${AZURE_TENANT_ID}`]
