+++ src/services/config/i18next.js
@@ -5,6 +5,7 @@ import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 import HttpApi from 'i18next-http-backend';
+import { LANGUAGES, FALLBACK_LANGUAGE } from '../../config/AppConfiguration';
 
 const langDetectorOptions = {
   // order and from where user language should be detected
@@ -29,8 +30,8 @@ i18n
   .use(initReactI18next)
 // for all options read: https://www.i18next.com/overview/configuration-options
   .init({
-    fallbackLng: 'en',
-    whitelist: ['en', 'fr'],
+    fallbackLng: FALLBACK_LANGUAGE,
+    whitelist: Object.keys(LANGUAGES),
     detection: langDetectorOptions,
     backend: {
       loadPath: '/locales/{{lng}}/{{ns}}.json'
