+++ src/state/sagas/scenario/CreateScenario/CreateScenarioData.js
@@ -4,9 +4,9 @@
 import { put, takeEvery, call } from 'redux-saga/effects';
 import { SCENARIO_ACTIONS_KEY } from '../../../commons/ScenarioConstants';
 import { STATUSES } from '../../../commons/Constants';
-import { ORGANIZATION_ID } from '../../../../configs/App.config';
+import { ORGANIZATION_ID } from '../../../../config/AppInstance';
 import { getAllScenariosData } from '../FindAllScenarios/FindAllScenariosData';
-import { Api } from '../../../../configs/Api.config';
+import { Api } from '../../../../services/config/Api';
 import { formatParametersFromApi } from '../../../../utils/ApiUtils';
 
 export function * createScenario (action) {
